+++ components/TextFields/tests/unit/ContainedInputView/MDCBaseTextFieldLayoutTests.m
@@ -39,7 +39,7 @@
                                                        isEditing:(BOOL)isEditing {
   CGSize textFieldSize = CGSizeMake(100, 30);
   UIFont *font = [UIFont systemFontOfSize:[UIFont systemFontSize]];
-  UIFont *floatingFont = [font fontWithSize:(font.pointSize * 0.5)];
+  UIFont *floatingFont = [font fontWithSize:(font.pointSize * (CGFloat)0.5)];
   MDCBaseTextFieldLayout *layout = [[MDCBaseTextFieldLayout alloc]
       initWithTextFieldSize:textFieldSize
        positioningReference:[MDCContainedInputViewVerticalPositioningGuideBase new]
